models/MAT/requirements.txt
albumentations==0.5.2
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
bash-kernel==0.9.3
beautifulsoup4==4.9.3
bleach==6.1.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
charset-normalizer==3.3.2
comm==0.2.1
conda==4.9.1
conda-build==3.20.5
conda-package-handling==1.7.0
contourpy==1.1.1
cryptography==2.9.2
cycler==0.12.1
dataclasses==0.6
debugpy==1.8.0
decorator==4.4.2
defusedxml==0.7.1
dnspython==2.0.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.0.12
fonttools==4.47.0
future==0.18.2
gdown==4.7.1
glob2==0.7
idna==2.9
imageio==2.33.1
imgaug==0.4.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
ipykernel==6.28.0
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.1
jedi==0.17.2
Jinja2==3.1.2
json5==0.9.14
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-archive==3.4.0
jupyter-client==8.6.0
jupyter-console==6.6.3
jupyter-core==5.6.1
jupyter-events==0.9.0
jupyter-http-over-ws==0.0.8
jupyter-lsp==2.2.1
jupyter-server==2.12.1
jupyter-server-terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-pygments==0.3.0
jupyterlab-server==2.25.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
kornia==0.4.1
lazy-loader==0.3
libarchive-c==2.9
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mistune==3.0.2
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nbzip==0.1.0
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.6
notebook-shim==0.2.3
numpy==1.20.3
olefile==0.46
opencv-python==4.4.0.46
opencv-python-headless==4.4.0.46
overrides==7.4.0
packaging==23.2
pandas==1.3.2
pandocfilters==1.5.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
pkginfo==1.6.0
pkgutil-resolve-name==1.3.10
platformdirs==4.1.0
pluggy==1.3.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.7.2
ptyprocess==0.6.0
pure-eval==0.2.2
pycosat==0.6.3
pycparser==2.20
Pygments==2.7.1
pyOpenSSL==19.1.0
pyparsing==3.1.1
PySocks==1.7.1
pytest==7.4.4
python-dateutil==2.8.2
python-etcd==0.4.5
python-json-logger==2.0.7
pytz==2020.1
PyWavelets==1.4.1
PyYAML==5.3.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.32.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.16.2
ruamel-yaml==0.15.87
scikit-image==0.21.0
scipy==1.10.1
Send2Trash==1.8.2
shapely==2.0.2
six==1.14.0
sniffio==1.3.0
soupsieve==2.0.1
stack-data==0.6.3
terminado==0.18.0
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
torch==1.7.0
torchelastic==0.2.1
torchvision==0.8.0
tornado==6.4
tqdm==4.46.0
traitlets==5.14.1
typing-extensions==4.9.0
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
zipp==3.17.0
